added config override per developer, via Spring Boot profiles.
See the example for myself (jbeutel), and tests.  This didn't work with HibernateSpec, which I'm not sure is ready for prime-time, so I changed VolunteerHibernateSpec to VolunteerIntegrationSpec.  GebConfig.groovy can't use this either, unfortunately, because Geb ignores the Grails/Spring config.  I also found some limitations in the functional spec, for which I added comments.