fixed my previous commit to consolidate nonce
for the view to use the new nonce from the model, not one from volunteer.

This also saves volunteer despite a failure from Facebook, re-displays unsaved volunteer values in the case of an error (although we'll be avoiding those on the server side in the future), and fixes the tests.