added opt-in callback and got recipientId.
The app still tries to send to Facebook after 1 second, even if it hasn’t gotten the optin yet, for the demo.  But, at least this give Facebook plenty of time to process the user's opt-in request.  The delay is 0, polling the db every 0.1 seconds for the optin.  I used a test library, Spock, for PollingConditions, because I don’t know offhand of a similar non-test library.

To get the optin callback, the Facebook app needs to be subscribed to it, by selecting the “messaging_optins” field on its webhooks config.

This saves the recipientId in the db, for sending messages in the future, as the nonce may only work once.  I changed the Volunteer.recipientId from Long to String, so this will require a db reset on Heroku.